refactor(macos): extract shared composer pill-menu component
