+++ clients/macos/vellum-assistant/Features/Chat/ComposerThresholdPicker.swift
@@ -129,14 +129,6 @@ struct ComposerThresholdPicker: View {
     /// GET responses don't overwrite an optimistic selection.
     @State private var hasUserInteracted: Bool = false
 
-    #if os(macOS)
-    @State private var isMenuOpen = false
-    @State private var activePanel: VMenuPanel?
-    @State private var triggerFrame: CGRect = .zero
-    #endif
-
-    private let composerActionButtonSize: CGFloat = 32
-
     private var pillLabelColor: Color {
         switch currentPreset {
         case .fullAccess: return VColor.systemNegativeStrong
@@ -147,39 +139,33 @@ struct ComposerThresholdPicker: View {
 
     var body: some View {
         #if os(macOS)
-        Button {
-            if isMenuOpen {
-                activePanel?.close()
-                activePanel = nil
-                isMenuOpen = false
-            } else {
-                showMenu()
-            }
-        } label: {
-            HStack(spacing: 4) {
-                VIconView(currentPreset.icon, size: 14)
-                    .foregroundStyle(currentPreset.iconColor)
-                Text(currentPreset.label)
-                    .font(VFont.labelDefault)
-                    .foregroundStyle(pillLabelColor)
-                VIconView(.chevronDown, size: 10)
-                    .foregroundStyle(VColor.contentTertiary)
-            }
-            .frame(height: composerActionButtonSize)
-            .padding(.horizontal, VSpacing.xs)
-            .contentShape(Rectangle())
-        }
-        .buttonStyle(.plain)
-        .vTooltip(currentPreset.description)
-        .accessibilityLabel("Risk tolerance")
-        .accessibilityValue(currentPreset.label)
-        .overlay {
-            GeometryReader { geo in
-                Color.clear
-                    .onAppear { triggerFrame = geo.frame(in: .global) }
-                    .onChange(of: geo.frame(in: .global)) { _, newFrame in
-                        triggerFrame = newFrame
+        ComposerPillMenu(
+            accessibilityLabel: "Risk tolerance",
+            accessibilityValue: currentPreset.label,
+            tooltip: currentPreset.description
+        ) {
+            VIconView(currentPreset.icon, size: 14)
+                .foregroundStyle(currentPreset.iconColor)
+            Text(currentPreset.label)
+                .font(VFont.labelDefault)
+                .foregroundStyle(pillLabelColor)
+        } menu: {
+            ForEach(ThresholdPreset.allCases) { preset in
+                VMenuItem(
+                    icon: preset.icon.rawValue,
+                    label: preset.label,
+                    isActive: currentPreset == preset,
+                    size: .regular
+                ) {
+                    selectPreset(preset)
+                } trailing: {
+                    VStack(alignment: .trailing, spacing: 2) {
+                        if currentPreset == preset {
+                            VIconView(.check, size: 12)
+                                .foregroundStyle(VColor.primaryBase)
+                        }
                     }
+                }
             }
         }
         .task(id: conversationId) {
@@ -188,68 +174,6 @@ struct ComposerThresholdPicker: View {
         }
         #endif
     }
-
-    // MARK: - Menu
-
-    #if os(macOS)
-    private func showMenu() {
-        guard !isMenuOpen else { return }
-        isMenuOpen = true
-
-        NSApp.keyWindow?.makeFirstResponder(nil)
-
-        guard let window = NSApp.keyWindow ?? NSApp.windows.first(where: { $0.isVisible }) else {
-            isMenuOpen = false
-            return
-        }
-
-        let triggerInWindow = CGPoint(x: triggerFrame.minX, y: triggerFrame.maxY)
-        let screenPoint = window.convertPoint(toScreen: NSPoint(
-            x: triggerInWindow.x,
-            y: window.frame.height - triggerInWindow.y
-        ))
-
-        let triggerScreenOrigin = window.convertPoint(toScreen: NSPoint(
-            x: triggerFrame.minX,
-            y: window.frame.height - triggerFrame.maxY
-        ))
-        let triggerScreenRect = CGRect(
-            origin: triggerScreenOrigin,
-            size: CGSize(width: triggerFrame.width, height: triggerFrame.height)
-        )
-
-        let appearance = window.effectiveAppearance
-        activePanel = VMenuPanel.show(
-            at: screenPoint,
-            sourceWindow: window,
-            sourceAppearance: appearance,
-            excludeRect: triggerScreenRect
-        ) {
-            VMenu(width: 240) {
-                ForEach(ThresholdPreset.allCases) { preset in
-                    VMenuItem(
-                        icon: preset.icon.rawValue,
-                        label: preset.label,
-                        isActive: currentPreset == preset,
-                        size: .regular
-                    ) {
-                        selectPreset(preset)
-                    } trailing: {
-                        VStack(alignment: .trailing, spacing: 2) {
-                            if currentPreset == preset {
-                                VIconView(.check, size: 12)
-                                    .foregroundStyle(VColor.primaryBase)
-                            }
-                        }
-                    }
-                }
-            }
-        } onDismiss: {
-            isMenuOpen = false
-            activePanel = nil
-        }
-    }
-    #endif
 
     // MARK: - Selection
